+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - A mocked Social Media indicator
 - Ability to choose from a predefined list of social media
 - EAS dev build config
+- Ability to vertically pan the image (not affecting slicing yet)
 
 ### Changed
 
@@ -41,6 +42,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated README to more accurately present the app
 - Moved gesture handling logic into a separate hook component
 - Refactored the project's folder structure
+- Changed Image Slider from using FlatList to ScrollView for gesture compatibility
 
 ### Removed